requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachecontrol==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.15
django-admin-geomap==1.6
django-cors-headers==4.4.0
django-filter==24.3
django-jet-reboot==1.3.9
django-js-asset==2.2.0
django-mptt==0.14.0
django-oauth-toolkit==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
firebase-admin==6.5.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.9
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
msgpack==1.1.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.3
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
